+++ CHANGES.rst
@@ -3,7 +3,11 @@ Changelog
 
 0.2.3 (unreleased)
 ------------------
-
+- Now pulls the unified installer [pigeonflight]
+    - downloads unified installer
+    - unpacks eggs cache
+    - unpacks downloads cache
+    - configures buildout dir to use downloads cache
 - Remove:
     - BUILDOUT_OPT